Standardized Activities
Implemented consistent activity behavior for the Drawer Menu.

Incidentally, standardizes toolbar style.
