+++ sign-lingual-app-android/app/src/main/java/com/example/signlingual/CameraConnectionFragment.java
