+++ sign-lingual-app-android/app/src/main/java/com/example/signlingual/LiveTranslation.java
@@ -1,37 +1,23 @@
 package com.example.signlingual;
 
 import androidx.annotation.NonNull;
-import androidx.appcompat.app.ActionBar;
-import androidx.appcompat.app.AppCompatActivity;
-import androidx.appcompat.widget.Toolbar;
-import androidx.core.view.GravityCompat;
-import androidx.drawerlayout.widget.DrawerLayout;
 
-import android.content.Intent;
-import android.graphics.Color;
 import android.os.Bundle;
 import android.util.Log;
-import android.view.MenuItem;
 import android.view.View;
 import android.widget.Button;
 import android.widget.TextView;
 
-import com.google.android.material.navigation.NavigationView;
-import com.google.firebase.analytics.FirebaseAnalytics;
 import com.google.firebase.database.DataSnapshot;
 import com.google.firebase.database.DatabaseError;
 import com.google.firebase.database.DatabaseReference;
 import com.google.firebase.database.FirebaseDatabase;
 import com.google.firebase.database.ValueEventListener;
 
-import org.w3c.dom.Text;
-
-public class LiveTranslation extends AppCompatActivity implements NavigationView.OnNavigationItemSelectedListener {
+public class LiveTranslation extends BaseActivity {
     String message="";
     TextView conversation;
     Button launch, stop;
-    private NavigationView navigationView;
-    private DrawerLayout drawerLayout;
     private boolean readable = false;
 
     @Override
@@ -125,60 +111,7 @@ public class LiveTranslation extends AppCompatActivity implements NavigationView
 
 
 
-    private void setupUI() {
-        drawerLayout = findViewById(R.id.drawer_layout);
-        navigationView = findViewById(R.id.navigationView);
-        navigationView.setNavigationItemSelectedListener(this);
-
-        Toolbar toolbar = findViewById(R.id.toolbar);
-        setSupportActionBar(toolbar);
-
-        toolbar.setBackgroundColor(Color.BLACK);
-        toolbar.setTitle("SignLingual");
-        toolbar.setTitleTextColor(Color.WHITE);
-
-        getSupportActionBar().setTitle("SignLingual");
-        getSupportActionBar().setDisplayHomeAsUpEnabled(true);
-    }
-
-    @Override
-    public boolean onOptionsItemSelected(@NonNull MenuItem item) {
-        int itemID = item.getItemId();
-        if (itemID == android.R.id.home) {
-            if(drawerLayout.isDrawerOpen(GravityCompat.START)) {
-                drawerLayout.closeDrawer(GravityCompat.START);
-//                Toast.makeText(MainActivity.this, "inside", Toast.LENGTH_SHORT).show();
-            } else {
-                drawerLayout.openDrawer(GravityCompat.START);
-//                Toast.makeText(MainActivity.this, "else", Toast.LENGTH_SHORT).show();
-            }
-        }
-        return true;
-    }
-
-    @Override
-    public boolean onNavigationItemSelected(@NonNull MenuItem item) {
-        int itemID = item.getItemId();
-        Bundle params = new Bundle();  // Create a bundle to hold parameters
-
-        if (itemID == R.id.navHome) {
-            params.putString("message", "User navigated to Home");
-            Intent intent = new Intent(getApplicationContext(), MainActivity.class);
-            startActivity(intent);
-        } else if (itemID == R.id.page1) {
-            params.putString("message", "User navigated to Page 1");
-            Intent intent = new Intent(getApplicationContext(), CameraActivity.class);
-            startActivity(intent);
-        } else if (itemID == R.id.settings) {
-            params.putString("message", "User navigated to Settings");
-            Intent intent = new Intent(getApplicationContext(), SettingsActivity.class);
-            startActivity(intent);
-        }else if (itemID == R.id.liveTranslation) {
-            params.putString("message", "User navigated to Live Translation");
-            Intent intent = new Intent(getApplicationContext(), LiveTranslation.class);
-            startActivity(intent);
-        }
-        drawerLayout.closeDrawer(GravityCompat.START);
-        return true;
+    protected void setupUI() {
+        super.setupUI();
     }
 }
