+++ sign-lingual-app-android/app/src/main/java/com/example/signlingual/SettingsActivity.java
@@ -4,18 +4,15 @@ import android.os.Bundle;
 import android.util.Log;
 
 import androidx.annotation.Nullable;
-import androidx.appcompat.app.AppCompatActivity;
-import androidx.appcompat.widget.Toolbar;
 
-public class SettingsActivity extends AppCompatActivity {
-
-    Toolbar toolbar;
+public class SettingsActivity extends BaseActivity {
 
     @Override
     protected void onCreate(@Nullable Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_settings);
 
+        setupUI();
         // Requires update for the comment to work in androidx. It replaces everything until marker.
         /*
         getSupportFragmentManager()
@@ -23,13 +20,6 @@ public class SettingsActivity extends AppCompatActivity {
             .replace(R.id.idFrameLayout, new SettingsFragments())
             .commit();
          */
-        Toolbar toolbar = findViewById(R.id.toolbar_settings);
-        setSupportActionBar(toolbar);
-        getSupportActionBar().setTitle("Settings");
-        getSupportActionBar().setDisplayHomeAsUpEnabled(true);
-        getSupportActionBar().setDisplayShowHomeEnabled(true);
-
-
         if (findViewById(R.id.frame_settings) != null) {
             if (savedInstanceState != null) {
                 Log.d("SettingActivity", "Saved Instance State Not Found");
@@ -39,5 +29,10 @@ public class SettingsActivity extends AppCompatActivity {
             Log.i("SettingsActivity", "Fragment found and initialized");
         }
         // Marker
+    }
+
+    @Override
+    protected void setupUI() {
+        super.setupUI();
     }
 }
