+++ sign-lingual-app-android/app/src/main/java/com/example/signlingual/StandaloneActivity.java
@@ -1,21 +1,8 @@
 package com.example.signlingual;
 
-import androidx.annotation.NonNull;
-import androidx.appcompat.app.AppCompatActivity;
-import androidx.appcompat.widget.Toolbar;
-import androidx.core.view.GravityCompat;
-import androidx.drawerlayout.widget.DrawerLayout;
-
-import android.content.Intent;
 import android.os.Bundle;
-import android.view.MenuItem;
-import android.view.View;
 
-import com.google.android.material.navigation.NavigationView;
-
-public class StandaloneActivity extends AppCompatActivity implements NavigationView.OnNavigationItemSelectedListener {
-    private NavigationView navigationView;
-    private DrawerLayout drawerLayout;
+public class StandaloneActivity extends BaseActivity {
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
@@ -24,60 +11,7 @@ public class StandaloneActivity extends AppCompatActivity implements NavigationV
         setupUI();
     }
 
-    private void setupUI() {
-        drawerLayout = findViewById(R.id.drawer_layout);
-        navigationView = findViewById(R.id.navigationView);
-        navigationView.setNavigationItemSelectedListener(this);
-
-
-        Toolbar toolbar = findViewById(R.id.toolbar);
-        setSupportActionBar(toolbar);
-        getSupportActionBar().setTitle("SignLingual");
-        getSupportActionBar().setDisplayHomeAsUpEnabled(true);
-    }
-
-    @Override
-    public boolean onOptionsItemSelected(@NonNull MenuItem item) {
-        int itemID = item.getItemId();
-        if (itemID == android.R.id.home) {
-            if(drawerLayout.isDrawerOpen(GravityCompat.START)) {
-                drawerLayout.closeDrawer(GravityCompat.START);
-//                Toast.makeText(MainActivity.this, "inside", Toast.LENGTH_SHORT).show();
-            } else {
-                drawerLayout.openDrawer(GravityCompat.START);
-//                Toast.makeText(MainActivity.this, "else", Toast.LENGTH_SHORT).show();
-            }
-        }
-        return true;
-    }
-
-    @Override
-    public boolean onNavigationItemSelected(@NonNull MenuItem item) {
-        int itemID = item.getItemId();
-        Bundle params = new Bundle();  // Create a bundle to hold parameters
-
-        if (itemID == R.id.navHome) {
-            params.putString("message", "User navigated to Home");
-            Intent intent = new Intent(getApplicationContext(), MainActivity.class);
-            startActivity(intent);
-        } else if (itemID == R.id.page1) {
-            params.putString("message", "User navigated to Page 1");
-            Intent intent = new Intent(getApplicationContext(), CameraActivity.class);
-            startActivity(intent);
-        } else if (itemID == R.id.settings) {
-            params.putString("message", "User navigated to Settings");
-            Intent intent = new Intent(getApplicationContext(), SettingsActivity.class);
-            startActivity(intent);
-        }else if (itemID == R.id.liveTranslation) {
-            params.putString("message", "User navigated to Live Translation");
-            Intent intent = new Intent(getApplicationContext(), LiveTranslation.class);
-            startActivity(intent);
-        } else if (itemID == R.id.standaloneMode) {
-            params.putString("message", "User navigated to Standalone");
-            Intent intent = new Intent(getApplicationContext(), StandaloneActivity.class);
-            startActivity(intent);
-        }
-        drawerLayout.closeDrawer(GravityCompat.START);
-        return true;
+    protected void setupUI() {
+        super.setupUI();
     }
 }
